Refactor MoreMenu to component
